rrweb-to-video/python/requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
ffmpeg-python>=0.2.0